// src/app/shared/models/limination/limination.model.ts
export interface Limination {
  id?: string;               
  serviceName: string;       
  price: number;             
  image?: string;            
  status: boolean;           
  description?: string;      
  createdAt?: any;           
  size?: string;             
  material?: string;         
  deliveryTime?: string;     
  discount?: number;         
  category?: string;         
  stockAvailable?: number;   
  categoryId: string;
  categoryName: string;
  
}